% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_508_table.R
\name{create_508_table}
\alias{create_508_table}
\title{Create a 508-compliant HTML table from tabular data}
\usage{
create_508_table(.data)
}
\description{
\code{create_508_table} returns an HTML string representing a 508-compliant version of the tabular data \code{.data}.
}
\examples{
df <- tibble('location'=c('Denver', 'DC', 'Miami'), 'meal'= c('dinner', 'lunch', 'lunch'))
html_text <- create_508_table(df)

}
