% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_508_table.R
\name{create_508_table}
\alias{create_508_table}
\title{Create a 508-compliant HTML table from tabular data}
\usage{
create_508_table(.data, table_id = "table", rownames.as.headers = FALSE)
}
\arguments{
\item{.data}{A data frame}

\item{table_id}{A string representing the \code{id} attribute of the
resulting \code{<table>} node}

\item{rownames.as.headers}{A boolean toggling whether to include row headers
based on the \code{rownames(.data)}}
}
\value{
xml2 node containing \code{<table>}
}
\description{
\code{create_508_table} returns an HTML string representing a 508-compliant
version of the tabular data \code{.data}.
}
