% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_508_table.R
\name{create_508_table}
\alias{create_508_table}
\title{Create a 508-compliant HTML table from tabular data}
\usage{
create_508_table(.data, table_id = "table", rownames_as_headers = FALSE,
  top_left_cell = NULL, summary_attr = NULL, caption = NULL)
}
\arguments{
\item{.data}{A data frame}

\item{table_id}{A string representing the \code{id} attribute of the
resulting \code{<table>} node}

\item{rownames_as_headers}{A boolean toggling whether to include row headers
based on the \code{rownames(.data)}}

\item{top_left_cell}{A string indicating value of top-left cell; used when
rownames_as_headers if TRUE}

\item{summary_attr}{A string used as the table's summary-attribute}

\item{caption}{A string used to create a \code{<caption>} tag}
}
\value{
xml2 node containing \code{<table>}
}
\description{
\code{create_508_table} returns an HTML string representing a 508-compliant
version of the tabular data \code{.data}.
}
